[B] Fix 'missing rollbar' error in non-composer mode
We decided to only 'suggest' rollbar/rollbar in composer.json, because
otherwise TYPO3 bombs when trying to load the extension in
non-composer-based installations.

This makes the extension compatible with both non-composer-mode and
composer-mode-based installations but requires that rollbar/rollbar is
included manually in addition to this plugin when using composer-mode.